feat: moved mocked earn feature flag state into mocks directory
